reformat: minor reformat remove: disabled threading feat: diagonal arrows

+++ modules/callbacks/callback_handlers.py
@@ -1,16 +1,47 @@
-"""Handles the commands"""
-from threading import Thread
+"""Handles the callbacks"""
 import os
-import textwrap
-
-from telegram import Update, ParseMode, InlineKeyboardButton, InlineKeyboardMarkup, InputMediaPhoto
+from telegram import Update, ParseMode
 from telegram.ext import CallbackContext
-
-from modules.various.utils import get_message_info, get_callback_info
+from modules.various.utils import get_callback_info
 from modules.various.photo_utils import generate_photo, build_bg_path, build_photo_path
-
 from modules.data.data_reader import read_md
 from modules.commands.command_handlers import STATE
+
+OFFSET_VALUES = {
+    'up': {
+        'x': 0,
+        'y': 50
+    },
+    'down': {
+        'x': 0,
+        'y': -50
+    },
+    'left': {
+        'x': 50,
+        'y': 0
+    },
+    'right': {
+        'x': -50,
+        'y': 0
+    },
+    'up-left': {
+        'x': 50,
+        'y': 50
+    },
+    'up-right': {
+        'x': -50,
+        'y': 50
+    },
+    'down-left': {
+        'x': 50,
+        'y': -50
+    },
+    'down-right': {
+        'x': -50,
+        'y': -50
+    },
+}
+
 
 def template_callback(update: Update, context: CallbackContext) -> int:
     """Handles the template callback
@@ -33,12 +64,6 @@ def template_callback(update: Update, context: CallbackContext) -> int:
                                   parse_mode=ParseMode.MARKDOWN_V2)
     return STATE['title']
 
-OFFSET_VALUES = {
-    'up': {'x': 0, 'y': 50},
-    'down': {'x': 0, 'y': -50},
-    'left': {'x': 50, 'y': 0},
-    'right': {'x': -50, 'y': 0}
-}
 
 def image_crop_callback(update: Update, context: CallbackContext) -> int:
     """Handles the image crop callback
@@ -55,7 +80,7 @@ def image_crop_callback(update: Update, context: CallbackContext) -> int:
 
     info = get_callback_info(update, context)
 
-    operation = info["query_data"].split(",")[1]
+    operation = info["query_data"][11:]
 
     if operation == 'reset':
         context.user_data['background_offset'] = {'x': 0, 'y': 0}
@@ -64,8 +89,9 @@ def image_crop_callback(update: Update, context: CallbackContext) -> int:
 
         info['bot'].edit_message_reply_markup(chat_id=info['chat_id'], message_id=info['message_id'], reply_markup=None)
 
-        os.remove(build_photo_path(sender_id)) 
-        os.remove(build_bg_path(sender_id)) 
+        if os.path.exists(build_bg_path(sender_id)):
+            os.remove(build_bg_path(sender_id))
+        os.remove(build_photo_path(sender_id))
 
         return STATE['end']
     else:
